+++ src/Enum/SortOrderTypeEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Enum;
+
+class SortOrderTypeEnum
+{
+    public const ASC = 'ASC';
+    public const DESC = 'DESC';
+}