+++ CHANGELOG.md
@@ -2,9 +2,9 @@ Changelog
 =========
 
 ##### Version 5.03.00 (WIP)
+ * add sonata admin dependency
  * add frontend page template render
  * add Ansible deployment process management
- * fix responsiveness issues
 
 ##### Version 5.02.00 (2021-07-28)
  * add frontend entities
@@ -23,3 +23,4 @@ To Do
 =====
 
  * decide better highlighted homepage aspect ratio covers
+ * fix responsiveness issues